ml-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

numpy==1.26.4
scikit-learn==1.2.2
joblib==1.2.0

pillow==10.2.0
pydantic==2.6.1

tensorflow==2.15.0
